chore: split facility utils into separate files
